TrustPing: seperate handlers by message type
Also changed1 method name to follow camelCase
